# documentation/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
...
paths:
...
  /v1/plandefinition:
    get:
      tags:
        - PlanDefinition
      operationId: getPlanDefinitions
      parameters:
        - name: statusesToInclude
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanDefinitionDto'
    put:
      tags:
        - PlanDefinition
      operationId: updatePlanDefinition
      parameters:
        - name: planDefinitionDto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PlanDefinitionDto'
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
    post:
      tags:
        - PlanDefinition
      summary: Create a new PlanDefinition.
      description: Create a PlanDefinition.
      operationId: createPlanDefinition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanDefinitionRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Error during creation of PlanDefinition.
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "201":
          description: Successful operation.
          headers:
            Location:
              description: URL pointing to the created PlanDefinition.
              style: simple
  /v1/plandefinition/{id}/retire:
    put:
      tags:
        - PlanDefinition
      operationId: retirePlanDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
...
  /v1/plandefinition/{id}:
    patch:
      tags:
        - PlanDefinition
      operationId: patchPlanDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPlanDefinitionRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
...
  /v1/plandefinition/{id}/used:
    get:
      tags:
        - PlanDefinition
      summary: Checks if the plandefinition is in use by any careplans
      description: Returns true if the plandefinition is in use by careplans and otherwise
        false if not
      operationId: isPlanDefinitionInUse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
...
  /v1/careplan/plandefinition:
    get:
      tags:
        - CarePlan
      operationId: getPlanDefinitions_1
      parameters:
        - name: statusesToInclude
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanDefinitionDto'
components:
  schemas:
    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errorCode:
          type: integer
          format: int32
        errorText:
          type: string
    EnableWhen:
      type: object
      properties:
        answer:
          $ref: '#/components/schemas/AnswerDto'
        operator:
          type: string
          enum:
            - EQUAL
            - GREATER_THAN
            - LESS_THAN
            - GREATER_OR_EQUAL
            - LESS_OR_EQUAL
    MeasurementTypeDto:
      type: object
      properties:
        system:
          type: string
        code:
          type: string
        display:
          type: string
    Option:
      type: object
      properties:
        option:
          type: string
        comment:
          type: string
    QuestionDto:
      type: object
      properties:
        linkId:
          type: string
        text:
          type: string
        abbreviation:
          type: string
        helperText:
          type: string
        required:
          type: boolean
        questionType:
          type: string
          enum:
            - CHOICE
            - INTEGER
            - QUANTITY
            - STRING
            - BOOLEAN
            - DISPLAY
            - GROUP
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        enableWhen:
          type: array
          items:
            $ref: '#/components/schemas/EnableWhen'
        thresholds:
          type: array
          items:
            $ref: '#/components/schemas/ThresholdDto'
        measurementType:
          $ref: '#/components/schemas/MeasurementTypeDto'
        subQuestions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionDto'
        parentQuestion:
          $ref: '#/components/schemas/QuestionDto'
        deprecated:
          type: boolean
    BaseDto:
      type: object
      properties:
        id:
          type: string
          description: Id of the resource
    QuestionnaireDto:
      allOf:
        - $ref: "#/components/schemas/BaseDto"
        - type: object
          properties:
            title:
              type: string
            status:
              type: string
            questions:
              type: array
              items:
                $ref: '#/components/schemas/QuestionDto'
            callToAction:
              $ref: '#/components/schemas/QuestionDto'
            version:
              type: string
            lastUpdated:
              type: string
              format: date-time
    ThresholdDto:
      type: object
      properties:
        questionId:
          type: string
        type:
          type: string
          enum:
            - NORMAL
            - ABNORMAL
            - CRITICAL
        valueBoolean:
          type: boolean
        valueQuantityLow:
          type: number
          format: double
        valueQuantityHigh:
          type: number
          format: double
        valueOption:
          type: string
    FrequencyDto:
      type: object
      properties:
        weekdays:
          type: array
          items:
            type: string
            enum:
              - MON
              - TUE
              - WED
              - THU
              - FRI
              - SAT
              - SUN
        timeOfDay:
          type: string
    PlanDefinitionDto:
      allOf:
        - $ref: "#/components/schemas/BaseDto"
        - type: object
          properties:
            name:
              type: string
            title:
              type: string
            status:
              type: string
            created:
              type: string
              format: date-time
            lastUpdated:
              type: string
              format: date-time
            questionnaires:
              type: array
              items:
                $ref: '#/components/schemas/QuestionnaireWrapperDto'
    QuestionnaireWrapperDto:
      type: object
      properties:
        questionnaire:
          $ref: '#/components/schemas/QuestionnaireDto'
        frequency:
          $ref: '#/components/schemas/FrequencyDto'
        satisfiedUntil:
          type: string
          format: date-time
        thresholds:
          type: array
          items:
            $ref: '#/components/schemas/ThresholdDto'
...
    AnswerDto:
      type: object
      properties:
        linkId:
          type: string
        value:
          type: string
        answerType:
          type: string
          enum:
            - INTEGER
            - STRING
            - BOOLEAN
            - QUANTITY
            - GROUP
        subAnswers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerDto'
...
    CreatePlanDefinitionRequest:
      required:
        - planDefinition
      type: object
      properties:
        planDefinition:
          $ref: '#/components/schemas/PlanDefinitionDto'
...
    PatchPlanDefinitionRequest:
      type: object
      properties:
        name:
          type: string
        questionnaireIds:
          type: array
          items:
            type: string
        thresholds:
          type: array
          items:
            $ref: '#/components/schemas/ThresholdDto'
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - RETIRED
...